+++ HISTORY.md
@@ -1,3 +1,9 @@
+unreleased
+==========
+
+  * Remove `parse(req)` and `parse(res)` signatures
+    - Use the `content-type` module for content type parsing
+
 0.3.0 / 2014-09-07
 ==================